django-admin/requirements.txt
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-split-settings==1.3.1
django==4.2.11
gunicorn==22.0.0
psycopg2-binary==2.9.9
pydantic_settings==2.6.0
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sentry-sdk[django]==2.17.0
whitenoise==6.3.0
